don't run lilo stuff if lilo is not the bootloader

+++ slackware/switchToGenericKernel.sh
@@ -5,19 +5,22 @@
 
 /usr/share/mkinitrd/mkinitrd_command_generator.sh | sh
 
-if [ -e ~/liloGenericEntry.sh ]; then
+if [ -e /etc/lilo.conf ]; then
+  ## clear out cruft
+  if [ -e ~/liloGenericEntry.sh ]; then
+    rm -v ~/liloGenericEntry.sh
+  fi
+  ## this is the magic
+  echo "/usr/share/mkinitrd/mkinitrd_command_generator.sh -l $(find /boot/ -name 'vmlinuz-generic-*' | tail -1)" > ~/liloGenericEntry.sh
+  ## check for duplicate entries
+  if [ -z "`grep $(cat ~/liloGenericEntry.sh | cut -d- -f4) /etc/lilo.conf`" ]; then
+    sh ~/liloGenericEntry.sh >> /etc/lilo.conf
+  fi
+  ## now we can run lilo
+  lilo -v
+  ## clean up
   rm -v ~/liloGenericEntry.sh
+else
+  echo "You aren't using lilo, update your bootloader."
 fi
-
-echo "/usr/share/mkinitrd/mkinitrd_command_generator.sh -l $(find /boot/ -name 'vmlinuz-generic-*' | tail -1)" > ~/liloGenericEntry.sh
-
-## check for duplicate entries
-if [ -z "`grep $(cat ~/liloGenericEntry.sh | cut -d- -f4) /etc/lilo.conf`" ]; then
-  sh ~/liloGenericEntry.sh >> /etc/lilo.conf
-fi
-
-lilo -v
-
-## clean up
-rm -v ~/liloGenericEntry.sh
